+++ CHANGELOG.md
@@ -19,6 +19,9 @@ BIDS sidecars.
 
 - Fixed the event names `NIFTI:Load` and `Dataframe:Load` to be title case for both parts of the name.
 
+- Fixed a naming bug in `calculateASLWorkload` where studies having multiple sessions did not correctly anticipate the
+  naming of .status files that should be created for the study.
+
 ### Added
 
 - Added the new BIDSDataGrid Module.
@@ -26,9 +29,20 @@ BIDS sidecars.
 - Added the `Shortcut:Register` and `Shortcut:Unregister` IPCMain events in order to have components register/unregister
   keyboard shortcuts. The corresponding frontend channel is called `shortcutTriggered`.
 
+- Added the `App:NotificationSound` IPCMain event. Some implementations added when running Process Studies section of
+  the application.
+
 ### Changed
 
+- The Import Module now gives some info feedback when a study finishes import, as long as the user did not terminate
+  the process.
+
+- Prior to the running of Structural or ASL modules, the BIDS2Legacy lock folders are deleted for each subject that is
+  present in the anticipated filepaths.
+
 - Minor removal of unnecessary imports and other sorts of code refactoring.
+
+- Moved the logic of determining ExploreASL version to its own function `getExploreASLVersion`.
 
 ---